tests/e2e/requirements.txt
# E2E Testing Dependencies
pytest-playwright==0.5.2
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.0.1
pytest-xdist==3.7.0  # For parallel test execution
allure-pytest==2.13.5  # For test reporting (optional)